pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-enum-dict"
version = "1.0.0"
description = "Enum和Dict管理工具，快速集成到FastAPI项目"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "MNVR Team", email = "dev@mnvr.com"}
]
keywords = ["fastapi", "enum", "dict", "crud", "scaffolding", "code-generator"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "jinja2>=3.0.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/mnvr/fastapi-enum-dict"
Documentation = "https://github.com/mnvr/fastapi-enum-dict#readme"
Repository = "https://github.com/mnvr/fastapi-enum-dict"

[project.scripts]
enum-dict = "fastapi_enum_dict.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["fastapi_enum_dict*"]

[tool.setuptools.package-data]
fastapi_enum_dict = ["templates/**/*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
